Prosseguimento do Projeto Integrador

+++ src/main/java/com/example/projetointegrador/repositories/CarteiraRepository.java
@@ -3,5 +3,5 @@ package com.example.projetointegrador.repositories;
 import com.example.projetointegrador.models.Carteira;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarteiraRepository extends JpaRepository<Carteira, Double> {
+public interface CarteiraRepository extends JpaRepository<Carteira, Long> {
 }
